Updated location and added more tests
