--- src/core/isTrue.js
@@ -1,9 +1,0 @@
-/** @license ISC License (c) copyright 2019 original and current authors */
-/** @author Dale Francis (dalefrancis88) */
-
-// isTrue : a -> Boolean
-function isTrue(x) {
-  return x === true
-}
-
-module.exports = isTrue
