+++ src/predicates/isFalse.js
@@ -1,5 +1,9 @@
-/** @license ISC License (c) copyright 2017 original and current authors */
+/** @license ISC License (c) copyright 2019 original and current authors */
 /** @author Dale Francis (dalefrancis88) */
 
-module.exports =
-  require('../core/isFalse')
+// isFalse : a -> Boolean
+function isFalse(x) {
+  return x === false
+}
+
+module.exports = isFalse
