+++ src/predicates/isTrue.js
@@ -1,5 +1,9 @@
 /** @license ISC License (c) copyright 2019 original and current authors */
 /** @author Dale Francis (dalefrancis88) */
 
-module.exports =
-  require('../core/isTrue')
+// isTrue : a -> Boolean
+function isTrue(x) {
+  return x === true
+}
+
+module.exports = isTrue
