+++ src/predicates/isTruthy.js
@@ -1,5 +1,9 @@
 /** @license ISC License (c) copyright 2019 original and current authors */
 /** @author Dale Francis (dalefrancis88) */
 
-module.exports =
-  require('../core/isTruthy')
+// isTruthy : a -> Boolean
+function isTruthy(x) {
+  return !!x
+}
+
+module.exports = isTruthy
